Define Kunshan's native-safe predicate
Show that it is a subset of the complement of the traced types, but note
in passing that because FuncRefs are not traced (correctly I think),
FuncRefs are neither traced nor native-safe.  Neither will composite
types containing them.

With FuncRef untraced, it is possible to have a Ptr (FuncRef _ _) and
for this value to be handed off to the "native world". I have no opinion
yet as to whether or not this is a good thing. I just report the facts,
ma'am.