Thread Semantics.
Still sketchy. Based on meeting with Eliot and Tony plus some subsequent
work on possible monad to be used to abstract and structure
specification. Not sure if we want to just "read" the threadstate (as in
Reader monad), or make it a state monad where we are always allowing for
threadstate transformation.  The memory messages are clearly a Writer
monad (a list of these that accumulates over the bind), and the handling
of errors (Blocked and Abort) is standard.